# openapi.yaml
openapi: 3.0.3
info:
  title: WarrantyVault API
  description: |
    REST API for native mobile clients (iOS / Android) of WarrantyVault.
    Source-of-truth for client codegen:
      - iOS:     swift-openapi-generator
      - Android: openapi-generator-cli (kotlin)
    The web app continues to use Next.js Server Actions — those are NOT
    documented here.

    Auth: bearer token from POST /api/v1/auth/login (or /register). Send as
    `Authorization: Bearer <token>` on every authenticated request.

    Ngôn ngữ: mọi chữ trong `message` / `fieldErrors` được server dịch sẵn. Thứ
    tự quyết định: `?lang=vi|en` → header `Accept-Language` → `User.locale` (đã
    lưu) → `en`. Ngôn ngữ không hỗ trợ rơi về mặc định, **không** phải lỗi 400.
    Xem `components.parameters.Lang`.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://warrantyvault.example.com
    description: Production (set when deployed)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/subscriptions:
    get:
      tags: [subscriptions]
      parameters:
        - name: status
          in: query
          description: |
            Filter by subscription status. Omit the parameter (or send an empty value)
            to get every status. Unlike `/api/v1/devices` there is no `ALL` sentinel:
            any other value is rejected with 400 `bad_input` and a Vietnamese field
            error on `status`.
          schema: { type: string, enum: [ACTIVE, PAUSED, CANCELED, EXPIRED] }
      responses:
        '200':
          description: Subscription list
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscriptions:
                    type: array
                    items: { $ref: '#/components/schemas/Subscription' }
        '400': { $ref: '#/components/responses/BadInput' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubscriptionInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties: { subscription: { $ref: '#/components/schemas/Subscription' } }
        '400': { $ref: '#/components/responses/BadInput' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/subscriptions/{id}:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    get:
      tags: [subscriptions]
      responses:
        '200':
          description: Subscription detail with payment history
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription: { $ref: '#/components/schemas/SubscriptionDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubscriptionInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties: { subscription: { $ref: '#/components/schemas/Subscription' } }
        '400': { $ref: '#/components/responses/BadInput' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [subscriptions]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
  /api/v1/subscriptions/{id}/payments:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    post:
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentInput' }
      responses:
        '201':
          description: Logged
          content:
            application/json:
              schema:
                type: object
                properties: { payment: { $ref: '#/components/schemas/Payment' } }
  /api/v1/subscriptions/{id}/renew:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    post:
      tags: [subscriptions]
      summary: Manually mark subscription as renewed (logs payment + advances renewalDate)
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/BadInput' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/subscriptions/audit:
    get:
      tags: [subscriptions]
      summary: Tự soát subscription — gói bị bỏ quên / tăng giá / trùng nhau (chỉ tư vấn)
      description: |
        Đọc **lịch sử thanh toán** để rút ra kết luận, thay vì chỉ hiển thị nó.
        Thuần SQL + Go, **không AI**, tất định và giải thích được từng dòng.

        ### ⚠️ Endpoint này KHÔNG sửa gì cả

        Không huỷ, không tắt `autoRenew`, không đổi giá, không ghi hàng nào.
        `advisory` luôn là `true`. Biến một phát hiện thành hành động là cú click
        của người dùng, qua các endpoint subscription đã có
        (`PATCH /api/v1/subscriptions/{id}`, `POST /api/v1/subscriptions/{id}/renew`).

        ### Trung thực về "dùng" — điều dữ liệu KHÔNG nói được

        App **không có tín hiệu sử dụng nào**: không telemetry, không đọc được
        giao dịch ngân hàng (roadmap §5 đã loại vì không có API tiêu dùng công
        khai), và schema không có gì ghi "dịch vụ này đã được mở". Vì vậy phát
        hiện **không** được gọi là "bạn không dùng gói này". Nó được gọi đúng
        tên: **"lâu rồi không thấy ghi nhận gì"**.

        Hai tín hiệu duy nhất tồn tại, và cả hai nói về việc **GHI NHẬN**, không
        phải việc dùng:
          * `SubscriptionPayment.note = 'Auto-renew'` — đúng chuỗi mà cron ghi
            khi tự trừ tiền (`internal/cron/run.go`). Tách được khoản máy trừ
            với khoản người dùng tự ghi.
          * các trường của `Subscription`, chỉ đổi khi người dùng sửa.

        ### Ba luật, và ngưỡng của chúng

        | `kind` | Điều kiện | Ngưỡng (hằng số trong Go) |
        |---|---|---|
        | `QUIET_AUTO_RENEW` | `ACTIVE` · `autoRenew = true` · không `LIFETIME` · **0 khoản do người dùng tự ghi** · số khoản máy tự trừ ≥ 3 · khoản tự trừ **đầu tiên** cách đây ≥ 6 tháng | `AuditQuietMinAutoCharges = 3`, `AuditQuietMinMonths = 6` |
        | `PRICE_INCREASED` | giá tăng giữa **hai kỳ thanh toán liền kề** (LAG trên toàn bộ lịch sử). Báo **mọi** mức tăng; `material` = `false` khi dưới 5% | `AuditPriceRiseMinPercent = 5` |
        | `DUPLICATE` | hai subscription `ACTIVE` cùng người dùng, **cùng tên sau khi bỏ dấu/viết thường** (`SAME_NAME`), hoặc **cùng hãng + cùng loại** (`SAME_BRAND_CATEGORY`). Loại trừ `LIFETIME` ở **cả hai** phía | — (không có ngưỡng) |

        * Đếm một mình sẽ gắn cờ gói `QUARTERLY` sau 9 tháng và gói `MONTHLY` sau
          3 tháng; yêu cầu thêm "khoản tự trừ đầu tiên ≥ 6 tháng" khiến luật có
          cùng ý nghĩa với mọi chu kỳ.
        * `AuditUpcomingRenewalDays = 14`: cửa sổ để một phát hiện còn hành động
          được trước khi tiền bị trừ; trả về trong `nextRenewalAt`/`daysUntilRenewal`.
        * So khớp trùng tên dùng `public.wv_unaccent` — đúng hàm mà tìm kiếm
          thiết bị đang dùng — nên "iCloud+" khớp "icloud+" và "Điện thoại" khớp
          "dien thoai" mà không cần thêm một bản chuẩn hoá thứ tư.
        * "Trùng nhau" ở đây nghĩa là **cùng đang hoạt động** (`ACTIVE`); schema
          không có ngày kết thúc cho một gói đang chạy, nên đó là định nghĩa
          "chồng lấn" duy nhất dữ liệu biểu diễn được.

        ### Vì sao tính ở server

        Phần **phát hiện** ở server, phần **hiển thị** ở client. Ba lý do cụ thể:
        (1) cần `wv_unaccent`, chỉ có trong Postgres; (2) cần `LAG` trên toàn bộ
        lịch sử thanh toán — làm ở client là N+1 query; (3) ngưỡng là quyết định
        sản phẩm, và ba client phải ra **cùng một lời khuyên** cho cùng một tài
        khoản. Giá trị ngưỡng được trả về trong `thresholds` để client hiển thị
        đúng luật đã áp dụng thay vì một phán quyết vô căn cứ.
      responses:
        '200':
          description: Báo cáo tự soát (chỉ tư vấn)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubscriptionAudit' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }

    # The language of the `message` / `fieldErrors` COPY in a response. It is not
    # a resource: it changes no status code, no error code and no field name, so a
    # client that ignores it loses nothing but the localisation.
    #
    # Precedence, first match wins:
    #   1. this parameter (`vi` | `en`)
    #   2. the `Accept-Language` request header (q-weighted, parsed by
    #      golang.org/x/text/language)
    #   3. the caller's stored preference (`User.locale`, set through PATCH
    #      /api/v1/auth/me)
    #   4. `en`
    #
    # A language we do not ship is a FALLBACK, never an error: `?lang=fr` is
    # ignored (the next level decides) and `Accept-Language: fr` gets English with
    # a 200. The response carries `Content-Language` when the request asked for a
    # language at all, so a client can see what it actually got.
...
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties: { ok: { type: boolean } }
...
    BadInput:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Missing / invalid / expired bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Resource not found (or owned by another user)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Conflict:
      description: Limit reached (per-user / per-resource cap exceeded)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string, description: 'Stable code, e.g. unauthorized, not_found, bad_input, rate_limited, limit_reached' }
        message: { type: string, description: 'Human-readable Vietnamese message' }
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    Subscription:
      type: object
      required: [id, name, billingCycle, price, startedAt, renewalDate, autoRenew, status]
      properties:
        id: { type: string }
        userId: { type: string }
        name: { type: string }
        category: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        plan: { type: string, nullable: true }
        billingCycle: { type: string, enum: [MONTHLY, QUARTERLY, YEARLY, LIFETIME, CUSTOM] }
        intervalDays: { type: integer, nullable: true }
        price: { type: integer, description: 'VND per cycle' }
        currency: { type: string, default: 'VND' }
        startedAt: { type: string, format: date-time }
        renewalDate: { type: string, format: date-time }
        autoRenew: { type: boolean }
        status: { type: string, enum: [ACTIVE, PAUSED, CANCELED, EXPIRED] }
        accountEmail: { type: string, nullable: true }
        paymentMethod: { type: string, nullable: true }
        manageUrl: { type: string, nullable: true }
        cancelUrl: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        lastNotifiedRenewalAt:
          type: string
          format: date-time
          nullable: true
          description: 'Set by the renewal-notification cron; null until the first renewal push goes out.'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SubscriptionDetail:
      allOf:
        - $ref: '#/components/schemas/Subscription'
        - type: object
          properties:
            payments:
              type: array
              items: { $ref: '#/components/schemas/Payment' }
    SubscriptionInput:
      type: object
      required: [name, billingCycle, price, startedAt, autoRenew]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        category: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        plan: { type: string, nullable: true }
        billingCycle: { type: string, enum: [MONTHLY, QUARTERLY, YEARLY, LIFETIME, CUSTOM] }
        intervalDays: { type: integer, minimum: 1, nullable: true, description: 'Required when billingCycle=CUSTOM' }
        price: { type: integer, minimum: 0 }
        startedAt: { type: string }
        renewalDate: { type: string, nullable: true, description: 'Defaults to nextRenewalDate(startedAt, billingCycle)' }
        autoRenew: { type: boolean }
        status: { type: string, enum: [ACTIVE, PAUSED, CANCELED, EXPIRED] }
        accountEmail: { type: string, nullable: true }
        paymentMethod: { type: string, nullable: true }
        manageUrl: { type: string, nullable: true }
        cancelUrl: { type: string, nullable: true }
        notes: { type: string, nullable: true }
    Payment:
      type: object
      required: [id, subscriptionId, amount, paidAt]
      properties:
        id: { type: string }
        subscriptionId: { type: string }
        amount: { type: integer }
        paidAt: { type: string, format: date-time }
        note: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    PaymentInput:
      type: object
      required: [amount, paidAt]
      properties:
        amount: { type: integer, minimum: 0 }
        paidAt: { type: string }
        note: { type: string, maxLength: 500, nullable: true }
...
    ActionCounts:
      type: object
      required: [total, high, medium, low]
      properties:
        total: { type: integer, description: 'Tổng số mục trong tập đang được đếm. Với `ActionQueue` đây là số việc **đang cần xử lý** (không tính việc bị hoãn, kể cả khi gọi `?snoozed=true`); với `SubscriptionAudit` là tổng số phát hiện.' }
        high: { type: integer }
        medium: { type: integer }
        low: { type: integer }
...
    SubscriptionAuditThresholds:
      type: object
      required: [quietMinAutoCharges, quietMinMonths, upcomingRenewalDays, priceRiseMinPercent, duplicateNormalized]
      properties:
        quietMinAutoCharges: { type: integer, description: 'Số khoản tự động trừ tối thiểu để xét `QUIET_AUTO_RENEW` (3).' }
        quietMinMonths: { type: integer, description: 'Khoản tự trừ ĐẦU TIÊN phải cách đây ít nhất bao nhiêu tháng (6).' }
        upcomingRenewalDays: { type: integer, description: 'Cửa sổ để một phát hiện còn hành động được trước khi bị trừ tiền (14 ngày).' }
        priceRiseMinPercent: { type: integer, description: 'Mức tăng (%) từ đó `material` = true. Mọi mức tăng vẫn được báo; ngưỡng này chỉ để client quyết định độ nổi bật (5).' }
        duplicateNormalized: { type: boolean, description: 'Luôn `true`: so khớp trùng tên có bỏ dấu và không phân biệt hoa/thường (`public.wv_unaccent`).' }
    SubscriptionAuditFinding:
      type: object
      required: [findingKey, kind, severity, title, detail, subscriptionIds, names, monthlyVnd, chargedTotalVnd, chargeCount]
      properties:
        findingKey: { type: string, description: '`<LOẠI_PHÁT_HIỆN>:<id hoặc cặp id>`. Ổn định giữa các lần gọi.' }
        kind: { type: string, enum: [QUIET_AUTO_RENEW, PRICE_INCREASED, DUPLICATE] }
        severity: { type: string, enum: [HIGH, MEDIUM, LOW] }
        title: { type: string }
        detail: { type: string, description: 'Câu tiếng Việt nêu đúng số liệu đã dẫn tới kết luận này.' }
        subscriptionIds: { type: array, items: { type: string }, description: 'Một id cho hai luật đầu, hai id cho `DUPLICATE`.' }
        names: { type: array, items: { type: string } }
        monthlyVnd: { type: integer, format: int64, description: 'Tổng quy đổi tháng của các gói liên quan (VND, int64). `0` với `LIFETIME`.' }
        chargedTotalVnd: { type: integer, format: int64, description: 'Tổng tiền **máy đã tự trừ** (`QUIET_AUTO_RENEW`). Không phải tổng mọi thanh toán.' }
        chargeCount: { type: integer, format: int64, description: 'Số khoản máy đã tự trừ.' }
        lastRecordedAt: { type: string, format: date-time, nullable: true, description: 'Ngày thanh toán mới nhất được ghi nhận. ⚠️ Đây **không** phải ngày dùng cuối — app không có dữ liệu đó.' }
        nextRenewalAt: { type: string, format: date-time, nullable: true }
        daysUntilRenewal: { type: integer, nullable: true }
        previousAmountVnd: { type: integer, format: int64, nullable: true, description: '`PRICE_INCREASED`: giá kỳ trước.' }
        amountVnd: { type: integer, format: int64, nullable: true, description: '`PRICE_INCREASED`: giá kỳ này.' }
        increaseVnd: { type: integer, format: int64, nullable: true }
        increasePercent: { type: integer, nullable: true, description: 'Mức tăng làm tròn theo %, `null` khi kỳ trước là 0 (không chia được).' }
        material: { type: boolean, nullable: true, description: '`false` khi mức tăng dưới `priceRiseMinPercent`. Phát hiện vẫn được trả về; chỉ độ nổi bật là do client quyết định.' }
        reason: { type: string, nullable: true, enum: [SAME_NAME, SAME_BRAND_CATEGORY], description: '`DUPLICATE`: tín hiệu nào đã khớp.' }
    SubscriptionAudit:
      type: object
      required: [generatedAt, findings, counts, advisory, thresholds, note]
      properties:
        generatedAt: { type: string, format: date-time }
        findings:
          type: array
          description: 'Luôn là mảng, không bao giờ null. Sắp xếp: `severity` giảm dần, rồi `monthlyVnd` giảm dần (tiền là lý do tính năng này tồn tại), rồi `findingKey`.'
          items: { $ref: '#/components/schemas/SubscriptionAuditFinding' }
        counts: { $ref: '#/components/schemas/ActionCounts' }
        advisory:
          type: boolean
          description: 'Luôn `true`. Có mặt để client khẳng định với người dùng rằng **không có gì bị sửa hay huỷ tự động** — endpoint này không có đường ghi nào.'
        thresholds: { $ref: '#/components/schemas/SubscriptionAuditThresholds' }
        note: { type: string, description: 'Câu tiếng Việt nói rõ giới hạn của phân tích (không đọc được giao dịch ngân hàng, không biết gói có đang được dùng hay không). Hiển thị thẳng.' }

    # ─── Warranty directory (FEATURE_IDEAS #15) ─────────────────────────────
...